refactor(frontend): tooltips
