+++ packages/frontend/components/Borrow/Overview/LTVProgressBar.tsx
@@ -10,7 +10,7 @@ import {
   Stack,
 } from "@mui/material"
 import InfoOutlinedIcon from "@mui/icons-material/InfoOutlined"
-import ClickableTooltip from "../../Shared/ClickableTooltip"
+import { ClickableTooltip } from "../../Shared/Tooltips"
 
 type LTVProgressBarProps = {
   borrowLimit: number
