+++ packages/frontend/components/Shared/Tooltips/index.ts
@@ -1,0 +1,2 @@
+export { default as ClickableTooltip } from "./ClickableTooltip"
+export { default as DocsTooltip } from "./DocsTooltip"
